Bugfixes and tweaks: Family, wizard robes, large cogwheels, dendor sacrifices (#2039)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

This PR fixes and tweaks the following:

- Wizard robes are now craftable, previously they had the same recipe
name as a wizard hat, making them unreachable to crafters
- Fixes several issues when examining family members, for example some
relations were missing or fathers saw their brothers as their sons and
so forth
- List family verb will now show real names of the family members, no
more "Unknown - Father" and such
- FamilyUI now defaults to ON if you are part of any family, no more
mothers not realizing they have a son and such
- Fixes opendream warning regarding dragon's attack
- Fixes crafting issues with large cogwheels
- Always call OnCrafted for all crafted items
- Fixes several issues with dendor sacrifices to prevent duplicit
activations

Fixes #662
Fixes #958
Fixes #2038
Fixes #1792
Fixes #1362

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Bugfixes.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

+++ code/modules/jobs/job_types/adventurer/types/pilgrim/briar.dm
@@ -168,19 +168,24 @@
 	layer = 4.2
 	alpha = 155
 	anchored = TRUE
+
 /obj/item/blessing_of_dendor_growing/attack_hand(mob/living/carbon/human/user)
 	if(user.patron.type == /datum/patron/divine/dendor)
 		icon_state = "dendor_grow_end"
+
+		if(!do_after(user, 3 SECONDS, target = user))
+			icon_state = "dendor_grow"
+			return
+
 		if(HAS_TRAIT(user, TRAIT_BLESSED))
 			to_chat(user, span_info("Dendor will not grant more powers, but he still approves of the sacrifice, judging by the signs..."))
 			user.apply_status_effect(/datum/status_effect/buff/blessed)
-			sleep(10)
 			GLOB.vanderlin_round_stats[STATS_DENDOR_SACRIFICES]++
 			qdel(src)
 			return
+
 		playsound(get_turf(user), 'sound/vo/smokedrag.ogg', 100, TRUE)
 		playsound(get_turf(user), 'sound/misc/wind.ogg', 100, TRUE, -1)
-		sleep(30)
 		to_chat(user, span_notice("Plants grow rampant with your every step...things that constrain no longer does."))
 		user.emote("smile")
 		ADD_TRAIT(user, TRAIT_BLESSED, TRAIT_GENERIC)
@@ -190,6 +195,7 @@
 			user.apply_status_effect(/datum/status_effect/buff/calm)
 	else
 		to_chat(user, span_warning("Dendor finds me unworthy..."))
+
 	GLOB.vanderlin_round_stats[STATS_DENDOR_SACRIFICES]++
 	qdel(src)
 
@@ -202,19 +208,24 @@
 	layer = 4.2
 	alpha = 155
 	anchored = TRUE
+
 /obj/item/blessing_of_dendor_stinging/attack_hand(mob/living/carbon/human/user)
 	if(user.patron.type == /datum/patron/divine/dendor)
 		icon_state = "dendor_sting_end"
+
+		if(!do_after(user, 3 SECONDS, target = user))
+			icon_state = "dendor_sting"
+			return
+
 		if(HAS_TRAIT(user, TRAIT_BLESSED))
 			to_chat(user, span_info("Dendor will not grant more powers, but he still approves of the sacrifice, judging by the signs..."))
 			user.apply_status_effect(/datum/status_effect/buff/blessed)
-			sleep(10)
 			GLOB.vanderlin_round_stats[STATS_DENDOR_SACRIFICES]++
 			qdel(src)
 			return
+
 		playsound(get_turf(user), 'sound/vo/smokedrag.ogg', 100, TRUE)
 		playsound(get_turf(user), 'sound/misc/wind.ogg', 100, TRUE, -1)
-		sleep(30)
 		to_chat(user, span_notice("You feel as if light follows your every step...your foraging will be easier from now on, surely."))
 		user.emote("smile")
 		ADD_TRAIT(user, TRAIT_BLESSED, TRAIT_GENERIC)
@@ -224,6 +235,7 @@
 			user.apply_status_effect(/datum/status_effect/buff/calm)
 	else
 		to_chat(user, span_warning("Dendor finds me unworthy..."))
+
 	GLOB.vanderlin_round_stats[STATS_DENDOR_SACRIFICES]++
 	qdel(src)
 
@@ -236,32 +248,41 @@
 	layer = 4.2
 	alpha = 155
 	anchored = TRUE
+
 /obj/item/blessing_of_dendor_devouring/attack_hand(mob/living/carbon/human/user)
 	if(user.patron.type == /datum/patron/divine/dendor)
 		icon_state = "dendor_consume_end"
+
+		if(!do_after(user, 3 SECONDS, target = user))
+			icon_state = "dendor_consume"
+			return
+
 		if(HAS_TRAIT(user, TRAIT_BLESSED))
 			to_chat(user, span_info("Dendor will not grant more powers, but he still approves of the sacrifice, judging by the signs..."))
 			user.apply_status_effect(/datum/status_effect/buff/blessed)
-			sleep(10)
 			GLOB.vanderlin_round_stats[STATS_DENDOR_SACRIFICES]++
 			qdel(src)
 			return
+
 		playsound(get_turf(user), 'sound/vo/smokedrag.ogg', 100, TRUE)
-		sleep(30)
 		to_chat(user, span_notice("A volf howls far away...and your teeth begin to sear with pain. Your sacrifice was accepted!"))
 		playsound(get_turf(user), 'sound/vo/mobs/wwolf/idle (1).ogg', 50, TRUE)
-		user.Immobilize(20)
-		sleep(20)
+		user.Immobilize(2 SECONDS)
+		sleep(2 SECONDS)
+
 		user.emote("pain")
-		sleep(5)
+		sleep(0.5 SECONDS)
+
 		playsound(get_turf(user), 'sound/combat/fracture/fracturewet (1).ogg', 70, TRUE, -1)
 		user.Immobilize(30)
-		sleep(35)
+		sleep(3.5 SECONDS)
+
 		to_chat(user, span_warning("My incisors transform to predatory fangs!"))
 		playsound(get_turf(user), 'sound/combat/fracture/fracturewet (1).ogg', 70, TRUE, -1)
 		user.emote("rage", forced = TRUE)
 		ADD_TRAIT(user, TRAIT_STRONGBITE, TRAIT_GENERIC)
 		ADD_TRAIT(user, TRAIT_BLESSED, TRAIT_GENERIC)
+
 		if(user.mind)
 			if(user.mind.has_spell(/obj/effect/proc_holder/spell/targeted/blesscrop))
 				user.apply_status_effect(/datum/status_effect/buff/barbrage)
@@ -273,5 +294,6 @@
 				to_chat(user, span_warning("I no longer care for mending wounds, let my rage be heard!"))
 	else
 		to_chat(user, span_warning("Dendor finds me unworthy..."))
+
 	GLOB.vanderlin_round_stats[STATS_DENDOR_SACRIFICES]++
 	qdel(src)
